Adding Swing form version of MoneyCalculator

+++ src/moneycalculator/model/Money.java
@@ -2,10 +2,10 @@ package moneycalculator.model;
 
 public class Money {
 
-        private final int amount;
+        private final double amount;
     private final Currency currency;
 
-    public Money(int amount, Currency currency) {
+    public Money(double amount, Currency currency) {
         this.amount = amount;
         this.currency = currency;
     }
